# server/swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: FABRIC Core API
  description: FABRIC Core API
  contact:
    name: Michael J. Stealey
    email: stealey@unc.edu
  license:
    name: MIT
    url: https://opensource.org/licenses/mit-license.html
  version: 1.0.0
servers:
- url: http://127.0.0.1:5000
  description: local development server (flask)
- url: https://127.0.0.1:8443
  description: local development server (nginx)
- url: https://virtserver.swaggerhub.com/mjstealey/fabric-core-api/1.0.0
  description: SwaggerHub API Auto Mocking
...
paths:
...
  /projects:
    get:
      tags:
      - projects
      summary: Search for FABRIC Projects
      description: Search for FABRIC Projects by name
      operationId: projects_get
      parameters:
      - name: search
        in: query
        description: search term applied
        required: false
        style: form
        explode: true
        schema:
          minLength: 3
          type: string
      - name: offset
        in: query
        description: number of items to skip before starting to collect the result
          set
        required: false
        style: form
        explode: true
        schema:
          minimum: 0
          type: integer
          format: int32
          default: 0
      - name: limit
        in: query
        description: maximum number of results to return per page (1 or more)
        required: false
        style: form
        explode: true
        schema:
          maximum: 20
          minimum: 1
          type: integer
          format: int32
          default: 5
      - name: person_uuid
        in: query
        description: person uuid
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/projects'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_400_bad_request'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_401_unauthorized'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_403_forbidden'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_404_not_found'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_500_internal_server_error'
      x-openapi-router-controller: swagger_server.controllers.projects_controller
    post:
      tags:
      - projects
      summary: Create new Project
      description: Create new Project
      operationId: projects_post
      requestBody:
        $ref: '#/components/requestBodies/payload_projects_post'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/projects_details'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_400_bad_request'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_401_unauthorized'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_403_forbidden'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_404_not_found'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_500_internal_server_error'
      x-openapi-router-controller: swagger_server.controllers.projects_controller
  /projects/{uuid}:
    get:
      tags:
      - projects
      summary: Project details by UUID
      description: Project details by UUID
      operationId: projects_uuid_get
      parameters:
      - name: uuid
        in: path
        description: universally unique identifier
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/projects_details'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_401_unauthorized'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_403_forbidden'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_404_not_found'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_500_internal_server_error'
      x-openapi-router-controller: swagger_server.controllers.projects_controller
    delete:
      tags:
      - projects
      summary: Delete Project as owner
      description: Delete Project as owner
      operationId: projects_uuid_delete
      parameters:
      - name: uuid
        in: path
        description: universally unique identifier
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_200_ok_no_content'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_401_unauthorized'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_403_forbidden'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_404_not_found'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_500_internal_server_error'
      x-openapi-router-controller: swagger_server.controllers.projects_controller
    patch:
      tags:
      - projects
      summary: Update Project details as owner
      description: Update Project details as owner
      operationId: projects_uuid_patch
      parameters:
      - name: uuid
        in: path
        description: universally unique identifier
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        $ref: '#/components/requestBodies/payload_projects_patch'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_200_ok_no_content'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_400_bad_request'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_401_unauthorized'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_403_forbidden'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_404_not_found'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_500_internal_server_error'
      x-openapi-router-controller: swagger_server.controllers.projects_controller
  /projects/{uuid}/profile:
    patch:
      tags:
      - projects
      summary: Update Project Profile details as owner
      description: Update Project Profile details as owner
      operationId: projects_uuid_profile_patch
      parameters:
      - name: uuid
        in: path
        description: universally unique identifier
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        $ref: '#/components/requestBodies/payload_projects_profile_patch'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_200_ok_no_content'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_400_bad_request'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_401_unauthorized'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_403_forbidden'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_404_not_found'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_500_internal_server_error'
      x-openapi-router-controller: swagger_server.controllers.projects_controller
  /projects/{uuid}/personnel:
    patch:
      tags:
      - projects
      summary: Update Project personnel as owner
      description: Update Project personnel as owner
      operationId: projects_uuid_personnel_patch
      parameters:
      - name: uuid
        in: path
        description: universally unique identifier
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        $ref: '#/components/requestBodies/payload_projects_personnel_patch'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_200_ok_no_content'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_400_bad_request'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_401_unauthorized'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_403_forbidden'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_404_not_found'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_500_internal_server_error'
      x-openapi-router-controller: swagger_server.controllers.projects_controller
  /projects/{uuid}/tags:
    patch:
      tags:
      - projects
      summary: Update Projects Tags as Facility Operator
      description: Update Projects Tags as Facility Operator
      operationId: projects_uuid_tags_patch
      parameters:
      - name: uuid
        in: path
        description: universally unique identifier
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        $ref: '#/components/requestBodies/payload_projects_tags_patch'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_200_ok_no_content'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_400_bad_request'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_401_unauthorized'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_403_forbidden'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_404_not_found'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_500_internal_server_error'
      x-openapi-router-controller: swagger_server.controllers.projects_controller
  /projects/preferences:
    get:
      tags:
      - projects
      summary: List of Projects Preference options
      description: List of Projects Preference options
      operationId: projects_preferences_get
      parameters:
      - name: search
        in: query
        description: search term applied
        required: false
        style: form
        explode: true
        schema:
          minLength: 3
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api_options'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_500_internal_server_error'
      x-openapi-router-controller: swagger_server.controllers.projects_controller
  /projects/profile/preferences:
    get:
      tags:
      - projects
      summary: List of Projects Profile Preference options
      description: List of Projects Profile Preference options
      operationId: projects_profile_preferences_get
      parameters:
      - name: search
        in: query
        description: search term applied
        required: false
        style: form
        explode: true
        schema:
          minLength: 3
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api_options'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_500_internal_server_error'
      x-openapi-router-controller: swagger_server.controllers.projects_controller
  /projects/tags:
    get:
      tags:
      - projects
      summary: List of Projects Tags options
      description: List of Projects Tags options
      operationId: projects_tags_get
      parameters:
      - name: search
        in: query
        description: search term applied
        required: false
        style: form
        explode: true
        schema:
          minLength: 3
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api_options'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status_500_internal_server_error'
      x-openapi-router-controller: swagger_server.controllers.projects_controller
...
components:
  schemas:
    status_200_ok_single:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
        size:
          type: integer
          default: 1
        status:
          type: integer
          default: 200
        type:
          type: string
    status_200_ok_paginated:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
        limit:
          type: integer
        links:
          $ref: '#/components/schemas/status_200_ok_paginated_links'
        offset:
          type: integer
        size:
          type: integer
        status:
          type: integer
          default: 200
        total:
          type: integer
        type:
          type: string
    status_200_ok_no_content:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/status_200_ok_no_content_data'
        type:
          type: string
          default: no_content
        size:
          type: integer
          default: 1
        status:
          type: integer
          default: 200
      example:
        data:
        - details: details
          message: No Content
        - details: details
          message: No Content
        size: 0
        type: no_content
        status: 6
    status_400_bad_request:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/status_400_bad_request_errors'
    status_401_unauthorized:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/status_401_unauthorized_errors'
        type:
          type: string
          default: error
        size:
          type: integer
          default: 1
        status:
          type: integer
          default: 401
    status_403_forbidden:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/status_403_forbidden_errors'
        type:
          type: string
          default: error
        size:
          type: integer
          default: 1
        status:
          type: integer
          default: 403
    status_404_not_found:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/status_404_not_found_errors'
        type:
          type: string
          default: error
        size:
          type: integer
          default: 1
        status:
          type: integer
          default: 404
    status_500_internal_server_error:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/status_500_internal_server_error_errors'
        type:
          type: string
          default: error
        size:
          type: integer
          default: 1
        status:
          type: integer
          default: 500
    person:
      required:
      - name
      - uuid
      type: object
      properties:
        email:
          type: string
        name:
          type: string
        uuid:
          type: string
...
    projects:
      type: object
      allOf:
      - $ref: '#/components/schemas/status_200_ok_paginated'
      - type: object
        properties:
          data:
            type: array
            items:
              $ref: '#/components/schemas/projects_data'
    projects_details:
      type: object
      allOf:
      - $ref: '#/components/schemas/status_200_ok_single'
      - type: object
        properties:
          data:
            type: array
            items:
              $ref: '#/components/schemas/projects_one'
    projects_one:
      required:
      - created
      - description
      - facility
      - is_public
      - name
      - uuid
      type: object
      properties:
        active:
          type: boolean
          default: true
        created:
          type: string
          format: date-time
        description:
          type: string
        facility:
          type: string
        is_public:
          type: boolean
          default: true
        memberships:
          $ref: '#/components/schemas/project_membership'
        modified:
          type: string
          format: date-time
        name:
          type: string
        preferences:
          $ref: '#/components/schemas/preferences'
        profile:
          $ref: '#/components/schemas/profile_projects'
        project_creators:
          type: array
          items:
            $ref: '#/components/schemas/person'
        project_members:
          type: array
          items:
            $ref: '#/components/schemas/person'
        project_owners:
          type: array
          items:
            $ref: '#/components/schemas/person'
        publications:
          type: array
          items:
            type: object
        tags:
          type: array
          items:
            type: string
        uuid:
          type: string
    project_membership:
      type: object
      properties:
        is_creator:
          type: boolean
          default: false
        is_member:
          type: boolean
          default: false
        is_owner:
          type: boolean
          default: false
...
    profile_projects:
      type: object
      properties:
        award_information:
          type: string
        goals:
          type: string
        keywords:
          type: array
          items:
            type: string
        notebooks:
          type: array
          items:
            type: object
        preferences:
          $ref: '#/components/schemas/preferences'
        project_status:
          type: string
        purpose:
          type: string
        references:
          type: array
          items:
            $ref: '#/components/schemas/profile_projects_references'
    preferences:
      type: object
      additionalProperties:
        type: boolean
      example:
        show_XYZ: true
...
    api_options:
      type: object
      allOf:
      - $ref: '#/components/schemas/status_200_ok_single'
      - type: object
        properties:
          data:
            type: array
            items:
              type: object
              additionalProperties:
                type: string
              example: show_XYZ
...
    projects_post:
      required:
      - description
      - is_public
      - name
      type: object
      properties:
        description:
          type: string
        is_public:
          type: boolean
          default: true
        name:
          type: string
        project_members:
          type: array
          items:
            type: string
        project_owners:
          type: array
          items:
            type: string
    projects_patch:
      type: object
      properties:
        description:
          type: string
        is_public:
          type: boolean
        name:
          type: string
        preferences:
          $ref: '#/components/schemas/preferences'
    projects_profile_patch:
      type: object
      properties:
        profile:
          $ref: '#/components/schemas/profile_projects'
    projects_personnel_patch:
      type: object
      properties:
        project_members:
          type: array
          items:
            type: string
        project_owners:
          type: array
          items:
            type: string
...
    status_200_ok_paginated_links:
      type: object
      properties:
        first:
          type: string
        last:
          type: string
        next:
          type: string
        prev:
          type: string
      example: null
    status_200_ok_no_content_data:
      type: object
      properties:
        message:
          type: string
          default: No Content
        details:
          type: string
      example:
        details: details
        message: No Content
    status_400_bad_request_errors:
      type: object
      properties:
        message:
          type: string
          default: Bad Request
        details:
          type: string
        type:
          type: string
          default: error
        size:
          type: integer
          default: 1
        status:
          type: integer
          default: 400
      example: null
    status_401_unauthorized_errors:
      type: object
      properties:
        message:
          type: string
          default: Unauthorized
        details:
          type: string
      example: null
    status_403_forbidden_errors:
      type: object
      properties:
        message:
          type: string
          default: Forbidden
        details:
          type: string
      example: null
    status_404_not_found_errors:
      type: object
      properties:
        message:
          type: string
          default: Not Found
        details:
          type: string
      example: null
    status_500_internal_server_error_errors:
      type: object
      properties:
        message:
          type: string
          default: Internal Server Error
        details:
          type: string
      example: null
...
    projects_data:
      required:
      - created
      - description
      - facility
      - is_public
      - name
      - uuid
      properties:
        created:
          type: string
        description:
          type: string
        facility:
          type: string
        is_public:
          type: boolean
          default: true
        memberships:
          $ref: '#/components/schemas/project_membership'
        name:
          type: string
        uuid:
          type: string
      example: null
...
    profile_projects_references:
      type: object
      properties:
        description:
          type: string
        url:
          type: string
      example: null
...
  requestBodies:
...
    payload_projects_post:
      description: Create new Project
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/projects_post'
    payload_projects_patch:
      description: Update Project details as owner
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/projects_patch'
    payload_projects_profile_patch:
      description: Update Project Profile details as owner
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/projects_profile_patch'
    payload_projects_personnel_patch:
      description: Update Project personnel as owner
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/projects_personnel_patch'
    payload_projects_tags_patch:
      description: Update Project Tags as Facility Operator
      content:
        application/json:
          schema:
            type: array
            items:
              type: string
...